requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.16
Pillow>=11.0

# Admin UI theme
django-jazzmin>=3.0

# API authentication
djangorestframework-simplejwt>=5.3

PyMySQL==1.1.1

# Useful extras
django-cors-headers>=4.4
django-environ>=0.11
# CREATE DATABASE motori CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci;

# CREATE USER 'mj'@'localhost' IDENTIFIED BY 'g100g100Aa$$';
# GRANT ALL PRIVILEGES ON motori.* TO 'mj'@'localhost';
# FLUSH PRIVILEGES;
# EXIT;
# GRANT SESSION_VARIABLES_ADMIN ON *.* TO 'mj'@'localhost';